Remove message handler logic
